small upd

+++ src/main/java/com/example/ws/microservices/firstmicroservices/repository/ShiftTimeWorkRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ShiftTimeWorkRepository extends JpaRepository<ShiftTimeWork, Integer> {
+public interface ShiftTimeWorkRepository extends JpaRepository<ShiftTimeWork, Short> {
 
     @Query(
             value = """
